+++ dpctl/tests/test_sycl_kernel_submit.py
@@ -244,7 +244,7 @@ def test_submit_async():
         e3_st = e3.execution_status
         e2_st = e2.execution_status
         e1_st = e1.execution_status
-        ht_e = q._submit_keep_args_alive([x_usm], [e1, e2, e3])
+        dpctl.keep_args_alive([x_usm], [e1, e2, e3])
         are_complete = [
             e == status_complete
             for e in (
@@ -254,7 +254,6 @@ def test_submit_async():
             )
         ]
         e3.wait()
-        ht_e.wait()
         if not all(are_complete):
             async_detected = True
             break